# [Java.Interop.Tools.JavaCallableWrappers] fix more places to use TypeDefinitionCache