… src/Java.Interop.Tools.TypeNameMappings/Java.Interop.Tools.TypeNameMappings/JavaNativeTypeManager.cs
… tools/generator/CodeGenerator.cs
… tools/generator/Java.Interop.Tools.Generator.Transformation/JavaTypeResolutionFixups.cs