Reviewing code in `JavaCallableWrapperGenerator`, I found places we weren't using the supplied `TypeDefinitionCache` or `IMetadataResolver`. Thus we appeared to be calling `TypeReference.Resolve()` potentially on the same types.

For example, `dotnet trace` output of an incremental build of a `dotnet new maui` project:

    41.65ms xamarin.android.cecil!Mono.Cecil.TypeReference.Resolve()

I created a new `TypeDefinitionRocks.ResolveCached()` method to be used everywhere instead. Resulting with:

    23.89ms xamarin.android.cecil!Mono.Cecil.TypeReference.Resolve()

Additionally, I updated to places to use plain `foreach` loops instead of System.Linq.

    Before:
    1.03s xamarin.android.build.tasks!Xamarin.Android.Tasks.GenerateJavaStubs.RunTask()
    After:
    944.48ms xamarin.android.build.tasks!Xamarin.Android.Tasks.GenerateJavaStubs.RunTask()

I think this likely saves about ~50ms off incremental builds of a `dotnet new maui` project.