@jonpryor 2023-01-04
At what point do we decide that `IMetadataResolver` is *required* and not optional?  I feel like we're playing "wack-a-mole" with callsites which *should* be providing a cache, but aren't, because we overlook them.

I think we should thus *require* `IMetadataResolver` instances, i.e. `IMetadataResolver resolver` instead of `IMetadataResolver? resolver`, and fix the fallout.