@jonpryor 2023-01-04
@jonpryor [wrote](https://github.com/xamarin/java.interop/pull/1069#issuecomment-1371275099):

> I think we should thus require `IMetadataResolver` instances, i.e. `IMetadataResolver resolver` instead of `IMetadataResolver? resolver`, and fix the fallout.

…and yes, this would be a partial "API break", in that currently non-warning code may start producing warnings, and `/warnaserror` would produce errors.  *That's the point*.  We *should* now require the caches, because we know it results in *significant* performance improvements.

This raises followup questions, such as what to do with our existing "compatibility overloads" such as https://github.com/xamarin/java.interop/blob/f8d77faf55347a58030a694332ba97f0dee88246/src/Java.Interop.Tools.Cecil/Java.Interop.Tools.Cecil/TypeDefinitionRocks.cs#L10-L12

We can't remove them; that would be an ABI break.

We should instead update them to be be an error to use:

```csharp
		[Obsolete ("Use the TypeDefinitionCache overload for better performance.", error:true)]
		public static TypeDefinition? GetBaseType (this TypeDefinition type) =>
			throw new NotSupportedException();
```